sealed_strategies/v3.3_20251216/locked/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quant-factor-trading"
version = "0.3.0"
description = "综合量化因子投资平台 - ETF轮动 + 因子库 + WFO优化"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Quant Team", email = "quant@example.com"}]
requires-python = ">=3.11"

# ============ 核心依赖 ============
dependencies = [
    # 数据处理（NumPy + Pandas + Polars）
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "polars>=1.33.1",
    # 科学计算
    "scipy>=1.16.2",
    "scikit-learn>=1.7.2",
    "statsmodels>=0.14.0",
    # 量化/技术分析
    "vectorbt>=0.28.1",
    "ta-lib>=0.6.7",
    "yfinance>=0.2.66",
    # 数据存储与序列化
    "pyarrow>=21.0.0",
    "fastparquet>=2024.11.0",
    # 配置与工具
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyscn>=1.1.1",
    # 网络与HTTP
    "requests>=2.32.0",
    "httpx[socks]>=0.28.1",
    # 并行与性能
    "joblib>=1.5.2",
    "numba>=0.62.0",
    # 可视化
    "matplotlib>=3.10.6",
    "seaborn>=0.13.2",
    "mplfinance>=0.12.10b0",
    "plotly>=5.24.0",
    # 代码质量
    "pandas-vet>=2023.8.2",
    "akshare>=1.17.91",
    "tushare>=1.4.0",
    "backtrader>=1.9.78.123",
]

# ============ 可选依赖 ============
[project.optional-dependencies]
# Web界面与API
web = [
    "dash>=2.18.0",
    "beautifulsoup4>=4.13.5",
    "flask>=3.0.0",
]

# 数据库与缓存
database = [
    "sqlalchemy>=2.0.0",
    "redis>=5.0.0",
]

# 任务调度
scheduling = [
    "schedule>=1.2.0",
]

# 开发与测试工具
dev = [
    # 单元测试
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    
    # 代码格式化与检查
    "black>=25.9.0",
    "isort>=6.0.1",
    "flake8>=7.3.0",
    "ruff>=0.13.2",
    "mypy>=1.18.2",
    
    # 安全扫描
    "bandit>=1.8.6",
    "safety>=3.6.2",
    
    # 性能分析
    "line-profiler>=5.0.0",
    "memory-profiler>=0.61.0",
    "vulture>=2.14",
    
    # Git钩子
    "pre-commit>=4.3.0",
    
    # 交互式开发
    "ipython>=9.5.0",
    "jupyterlab>=4.4.9",
    "notebook>=7.3.0",
    
    # 类型检查
    "pandas-stubs>=2.3.2.250926",
    "types-psutil>=7.0.0.20251001",
    "types-pyyaml>=6.0.12.20250915",
    "types-seaborn>=0.13.2.20250914",
]

# 所有可选依赖
all = [
    "quant-factor-trading[web,database,scheduling,dev]",
]

# ============ 工具配置 ============
[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | _archive
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
skip_glob = ["*/migrations/*", ".venv/*", "_archive/*"]

[tool.ruff]
line-length = 88
target-version = "py310"
exclude = [".venv", "_archive", "build", "dist", ".git"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["-v", "--strict-markers", "--tb=short"]
markers = [
    "slow: 标记为慢速测试",
    "integration: 标记为集成测试",
    "unit: 标记为单元测试",
]

[tool.coverage.run]
source = ["etf_rotation_optimized", "a_shares_strategy", "etf_download_manager"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

# ============ 项目URLs ============
[project.urls]
Homepage = "https://github.com/zhangsensen/quant-factor-trading"
Repository = "https://github.com/zhangsensen/quant-factor-trading.git"
Documentation = "https://github.com/zhangsensen/quant-factor-trading/blob/main/README.md"
"Bug Tracker" = "https://github.com/zhangsensen/quant-factor-trading/issues"
Changelog = "https://github.com/zhangsensen/quant-factor-trading/releases"

[tool.setuptools.packages.find]
where = ["src"]
include = ["etf_strategy*", "etf_data*"]
exclude = ["tests*", "scripts*", "docs*", "config*", "configs*", "raw*", "results*", "untitled*"]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "ruff>=0.14.5",
]
